+++ R/class_mrgmod.R
@@ -143,7 +143,9 @@ slots <- sapply(protomod, class)
 names(slots) <- names(protomod)
 
 valid.mrgmod <- function(object) {
-  tags <- unlist(names(object), use.names=FALSE)
+  tags <- names(object)
+  tags[["capture"]] <- NULL
+  tags <- unlist(tags, use.names=FALSE)
   x <- check_names(tags,Pars(object),Cmt(object))
   x1 <- length(x)==0
   x2 <- object@advan %in% c(0,1,2,3,4,13)
@@ -368,6 +370,7 @@ setMethod("names", "mrgmod", function(x) {
   ans <- list()
   ans$param <- Pars(x)
   ans$init <- Cmt(x)
+  ans$capture <- unname(x@capture)
   ans$omega <- names(omat(x))
   ans$sigma <- names(smat(x))
   ans$omega_labels <- labels(omat(x))
@@ -597,6 +600,31 @@ setMethod("blocks", "character", function(x,...) {
   what <- as.character(match.call()[-1])[-1]
   blocks_(x,what)
 })
+
+prvec <- function(x, ...) {
+  x <- strwrap(paste0(x,collapse=", "),...)    
+  paste0(x,collapse="\n")
+}
+
+#' Print summary of a mrgmod object
+#' @param object a mrgmod object
+#' @param ... not used
+#' @export
+summary.mrgmod <- function(object,...) {
+  l <- as.list(object)
+  ncmt <- l[["neq"]]
+  npar <- l[["npar"]]
+  message("Model: ", l$model)
+  message("- Parameters: [", npar, "]")
+  message(prvec(l$pars,width = 50,prefix="  "))
+  message("- Compartments: [", ncmt, "]")
+  message(prvec(l$cmt,width = 50, prefix = "  "))
+  message("- Captured: [", length(l[["capture"]]), "]")
+  o <- l$capture
+  if(length(o)==0) o <- "<none>"
+  message(prvec(o, width = 50, prefix = "  "))
+  return(invisible(NULL))
+}
 
 blocks_ <- function(file,what) {
   if(length(what)==0) what <- c("PARAM","MAIN", "ODE","DES", "TABLE")