+++ R/mrgsim_q.R
@@ -37,6 +37,9 @@
 ##' @param skip_init_calc don't use \code{$MAIN} to calculate initial conditions
 ##' @param output output data type; if \code{NULL}, then an \code{mrgsims}
 ##' object is returned; if \code{"df"} then a data frame is returned
+##' @param simcall the default value \code{1} will invoke use of a streamlined
+##' version of \code{DEVTRAN} to be use for the simulation; \code{0} revets
+##' back to the same \code{DEVTRAN} code that is used by \code{\link{mrgsim}}
 ##' 
 ##' @details
 ##' 
@@ -91,7 +94,7 @@ mrgsim_q <- function(x,
                      stime = numeric(0),
                      output = NULL,
                      skip_init_calc = FALSE, 
-                     simcall = 1) {
+                     simcall = 0) {
   
   ## data
   if(!is.valid_data_set(data)) {