+++ R/utils.R
@@ -596,3 +596,31 @@ reg_exec_match <- function(x, pattern) {
   regmatches(x,m)
 }
 
+collect_opts <- function(x) {
+  un <- unique(names(x))
+  ans <- lapply(un, function(y) {
+    unlist(x[names(x)==y],use.names=FALSE)
+  })
+  names(ans) <- un
+  ans
+}
+
+make_matrix_labels <- function(mat,lab,diag=TRUE) {
+  n <- nrow(mat)
+  cmat <- matrix(NA_character_, n, n)
+  for(i in seq(n)) {
+    for(j in seq(n)) {
+      if(i > j) next
+      if(i==j) {
+        val <- lab[i] 
+      } else {
+        val <- paste0(c(lab[i],lab[j]),collapse='-')
+      }
+      cmat[i,j] <- val
+    }
+  }
+  ans <- mat[upper.tri(mat,diag=diag)]
+  lab <- cmat[upper.tri(cmat,diag = diag)]
+  names(ans) <- lab
+  ans
+}