+++ NEWS.md
@@ -8,6 +8,13 @@ whether or not the streamlined function will be retained
 spaces were not handled properly #450
 - Objects of class `valid_data_set` now retain the matrix attribute #448
 - Block names are now coreced to upper case letters
+- Change default for simcall to 0 in mrgsim_q
+- Added recover argument to mread
+- Add `@code` option to the following blocks: PARAM, INIT, CMT, 
+  OMEGA, SIGMA
+- Allow multiple `ODE` blocks
+- Add `mtime` function to `self` object for simpler mtime implementation
+
 
 # mrgsolve 0.9.0
 - Added `logy` and `logbr` arguments to `plot.mrgsims` so results y-axis